Initial commit of mapbox_vite
mapbox_vite is a react application that is a sandbox to explore the
use of mapbox and vite in a react application. It currently consumes
and displays data from spatial_mus - a django sandbox build to explore
dynamic management units and different ways of selecting regions of
interest (by polygon and buffered point).